REFACTOR Cost, added functionality - Cost changed to Decimal - added ability to display a customer's order history - more functionality for repository

+++ MyStore.Store/IItem.cs
@@ -7,7 +7,7 @@ namespace MyStore.Store
 {
     public interface IItem : ISerializable
     {
-        public float cost { get; }
+        public decimal cost { get; }
         public string name { get;  }
 
         //private string UPCCode
